split of modules
